Fixing all the shit ive messed up
