+++ models/MatchTeam.js
@@ -1,0 +1,9 @@
+class MatchTeam {
+    constructor(teamNumber, tableId) {
+        this.teamNumber = teamNumber;
+        this.tableId = tableId;
+    }
+
+}
+
+module.exports = MatchTeam;
