+++ models/Table.js
@@ -1,0 +1,8 @@
+class Table {
+    constructor(tableId, tableName) {
+        this.tableId = tableId;
+        this.tableName = tableName;
+    }
+}
+
+module.exports = Table
