+++ routers/teamsRouter.js
@@ -4,12 +4,11 @@ const router = express.Router();
 const MongoClient = require('mongodb').MongoClient;
 
 const MONGU_URI = process.env.MONGO;
-const DB_NAME = 'test';
 
 
 router.get('/', (req, res) => {
     MongoClient.connect(MONGU_URI).then((connection) => {
-        connection.db(DB_NAME).collection('teams').find().toArray().then((data) => {
+        connection.db.collection('teams').find().toArray().then((data) => {
             res.send(data);
         });
     }).catch((err) => {
@@ -21,7 +20,7 @@ router.get('/', (req, res) => {
 
 router.get('/:id', (req, res) => {
     MongoClient.connect(MONGU_URI).then((connection) => {
-        connection.db(DB_NAME).collection('teams').findOne({'number': parseInt(req.params.id)}).then((data) => {
+        connection.db.collection('teams').findOne({'number': parseInt(req.params.id)}).then((data) => {
             if (!data) {
                 res.sendStatus(404);
                 return;
@@ -39,7 +38,7 @@ router.get('/:id', (req, res) => {
 
 router.put('/:id', (req, res) => {
     MongoClient.connect(MONGU_URI).then((connection) => {
-        connection.db(DB_NAME).collection('teams').findOneAndUpdate({'number': parseInt(req.params.id)}, req.body).then((a) => {
+        connection.db.collection('teams').findOneAndUpdate({'number': parseInt(req.params.id)}, req.body).then((a) => {
             if (a.ok === 1) {
                 res.sendStatus(200);
                 return;
@@ -55,13 +54,13 @@ router.put('/:id', (req, res) => {
 
 router.post('/', (req, res) => {
     MongoClient.connect(MONGU_URI).then((connection) => {
-        connection.db(DB_NAME).collection('teams').findOne({'number': parseInt(req.body.number)}).then((data) => {
+        connection.db.collection('teams').findOne({'number': parseInt(req.body.number)}).then((data) => {
             if (data) {
                 res.status(400);
                 res.send("Team allready exists");
                 return;
             }
-            connection.db(DB_NAME).collection('teams').insertOne(req.body).then((a) => {
+            connection.db.collection('teams').insertOne(req.body).then((a) => {
                 if (a.insertedCount > 0) {
                     res.sendStatus(200);
                     return;
@@ -78,7 +77,7 @@ router.post('/', (req, res) => {
 
 router.delete('/:id', (req, res) => {
     MongoClient.connect(MONGU_URI).then((connection) => {
-        connection.db(DB_NAME).collection('teams').findOneAndDelete({'number': parseInt(req.params.id)}).then((a) => {
+        connection.db.collection('teams').findOneAndDelete({'number': parseInt(req.params.id)}).then((a) => {
             if (a.ok === 1) {
                 res.sendStatus(200);
                 return;
